// src/main/conversation/dto/events.enum.ts
export enum ConversationEvents {
  SEND_MESSAGE = 'conversation:message_send',
  RECEIVE_MESSAGE = 'conversation:message_receive',
  MESSAGE_SENT = 'conversation:message_sent',
  MESSAGE_READ = 'conversation:message_read',
  TYPING_START = 'conversation:typing_start',
  TYPING_STOP = 'conversation:typing_stop',
  USER_STATUS = 'conversation:user_status',
  USER_STATUS_CHANGED = 'conversation:user_status_changed',
  GET_USER_STATUS = 'conversation:get_user_status',
  SET_USER_STATUS = 'conversation:set_user_status',
  LOAD_CONVERSATIONS = 'conversation:load_conversations',
  CONVERSATION_LIST = 'conversation:conversation_list',
  LOAD_SINGLE_CONVERSATION = 'conversation:load_single_conversation',
  CONVERSATION_MESSAGES = 'conversation:conversation_messages',
  CONVERSATION_USER_INFO = 'conversation:user_info',
  MESSAGE_REACT = 'conversation:message_react',
  MESSAGE_REACTED = 'conversation:message_reacted',
  MESSAGE_UNSEND = 'conversation:message_unsend',
  MESSAGE_UNSENT = 'conversation:message_unsent',
  MESSAGE_DELETED_FOR_ME = 'conversation:message_deleted_for_me',
  CONVERSATION_UPDATED = 'conversation:conversation_updated',
}